chore(web): Move the error-tracking utils to the observability/providers/mixpanel folder
