+++ apps/web/src/services/observability/providers/mixpanel/index.ts
@@ -1,0 +1,1 @@
+export { MixpanelTracingProvider } from './provider'
